Create `CONTRIBUTING.md`
This commit introduces the initial draft of our contribution guide.

The plan is to keep updating it as the project evolves, addressing any
commonly asked questions and providing additional clarification as
needed.

The guide is inspired by the excellent contribution guide from Hack For
LA. It has been heavily simplified to better suit the goals of our
project of accessibility, by attempting to not overwhelm new developers
with too much information.